Moved tests for `is_set` from `test_util.py` to `test_helpers.py`
