+++ asdf/_tests/test_util.py
@@ -7,7 +7,6 @@ import pytest
 import asdf
 import asdf.tagged
 from asdf import generic_io, util
-from asdf._helpers import _IsSet, is_set
 
 
 def test_not_set():
@@ -151,50 +150,3 @@ l: &id002 !some/tag-1.0.0
     tree = util.load_yaml(io.BytesIO(contents), tagged=tagged)
     assert tree["o"] is tree["o"]["inverse"]["inverse"]
     assert tree["l"] is tree["l"][0]
-
-
-class Tracked(_IsSet):
-    bar = None
-
-    def __init__(self):
-        # Intentionally not calling `super().__init__()` here to make sure `_IsSet` still works
-        self._value = 1
-        self.baz = None
-
-    @property
-    def value(self):
-        return self._value
-
-    @value.setter
-    def value(self, value):
-        self._value = value
-
-
-def test_is_set():
-    x = Tracked()
-    y = Tracked()
-
-    def is_set_iter(obj):
-        yield from (is_set(obj, attr) for attr in ["value", "bar", "baz"])
-
-    assert not any(is_set_iter(x))
-
-    x.value = 2
-    x.bar = 3
-    x.baz = "foo"
-
-    assert all(is_set_iter(x))
-    assert not any(is_set_iter(y))
-
-
-def test_is_set_attr_error():
-    """Test that trying to access a non-existent property via `is_set` raises an `AttributeError`."""
-    x = Tracked()
-    with pytest.raises(AttributeError):
-        is_set(x, "foo")
-
-
-def test_is_set_type_error():
-    with pytest.raises(TypeError):
-        # pyrefly: ignore [bad-argument-type]
-        is_set(object(), "foo")
